examples/zero-code-examples/openai-agents/requirements.txt
openai-agents>=0.3.3
opentelemetry-instrumentation-openai-agents-v2>=0.1.0

opentelemetry-sdk>=1.36.0
opentelemetry-exporter-otlp-proto-http>=1.36.0
python-dotenv>=1.0.0
